Extract link tooltip from editable text
REDMINE-20673
