+++ entry_types/scrolled/package/src/frontend/inlineEditing/LinkTooltip.js
@@ -1,19 +1,18 @@
 import React, {useContext, useState, createContext, useMemo, useRef} from 'react';
 import classNames from 'classnames';
-import {Range} from 'slate';
 
-import {useI18n} from '../../i18n';
-import {useChapter, useFile} from '../../../entryState';
-import {SectionThumbnail} from '../../SectionThumbnail';
+import {useI18n} from '../i18n';
+import {useChapter, useFile} from '../../entryState';
+import {SectionThumbnail} from '../SectionThumbnail';
 
-import styles from './index.module.css';
+import styles from './LinkTooltip.module.css';
 
-import ExternalLinkIcon from '../images/externalLink.svg';
+import ExternalLinkIcon from './images/externalLink.svg';
 
 const UpdateContext = createContext();
 
 export function LinkTooltipProvider({
-  editor, disabled, position, children, align = 'left', gap = 10
+  disabled, position, children, align = 'left', gap = 10
 }) {
   const [state, setState] = useState();
   const outerRef = useRef();
@@ -62,8 +61,7 @@ export function LinkTooltipProvider({
   return (
     <UpdateContext.Provider value={update}>
       <div ref={outerRef}>
-        <LinkTooltip editor={editor}
-                     state={state}
+        <LinkTooltip state={state}
                      disabled={disabled}
                      position={position}
                      align={align} />
@@ -86,21 +84,20 @@ export function LinkPreview({href, openInNewTab, children}) {
   );
 }
 
-export function LinkTooltip({editor, disabled, position, align, state}) {
+export function LinkTooltip({disabled, position, align, state}) {
   const {keep, deactivate} = useContext(UpdateContext);
 
-  if (disabled || !state || (editor.selection && !Range.isCollapsed(editor.selection))) {
+  if (disabled || !state) {
     return null;
   }
 
   return (
     <div className={classNames(styles.linkTooltip,
-                               styles[`linkTooltip-${position}`],
-                               styles[`linkTooltip-${align}`],
-                               styles.hoveringToolbar)}
+                               styles[`position-${position}`],
+                               styles[`align-${align}`])}
          onMouseEnter={keep}
          onMouseLeave={deactivate}
-         style={{top: state.top, bottom: state.bottom, left: state.left, opacity: 1}}>
+         style={{top: state.top, bottom: state.bottom, left: state.left}}>
       <LinkDestination href={state.href} openInNewTab={state.openInNewTab} />
     </div>
   );
@@ -142,7 +139,7 @@ function ChapterLinkDestination({permaId}) {
   return (
     <a href={`#${chapter.chapterSlug}`}
        title={t('pageflow_scrolled.inline_editing.link_tooltip.visit_chapter')}>
-      <span className={styles.linkTooltipChapterNumber}>
+      <span className={styles.chapterNumber}>
         {t('pageflow_scrolled.inline_editing.link_tooltip.chapter_number',
            {number: chapter.index + 1})}
       </span> {chapter.title}
@@ -154,10 +151,10 @@ function SectionLinkDestination({permaId}) {
   const {t} = useI18n({locale: 'ui'});
 
   return (
-    <div className={styles.linkTooltipThumbnail}>
+    <div className={styles.thumbnail}>
       <SectionThumbnail sectionPermaId={permaId} />
       <a href={`#section-${permaId}`}
-         className={styles.linkTooltipThumbnailClickMask}
+         className={styles.thumbnailClickMask}
          title={t('pageflow_scrolled.inline_editing.link_tooltip.visit_section')}>
       </a>
     </div>
@@ -175,7 +172,7 @@ function ExternalLinkDestination({href, openInNewTab}) {
         {href}
         <ExternalLinkIcon width={10} height={10} />
       </a>
-      <div className={styles.linkTooltipNewTab}>
+      <div className={styles.newTab}>
         {openInNewTab ?
          t('pageflow_scrolled.inline_editing.link_tooltip.opens_in_new_tab') :
          t('pageflow_scrolled.inline_editing.link_tooltip.opens_in_same_tab')}
